Whew!  Changing the structure of polytope a bit -- now requiring all tessellators to just provide one method that returns the QuantizedTessellation, and moving implementations for everything else (like clipping with PLC's) to the base class.  Only BoostTessellator working right now.

+++ src/KeyTraits.hh
@@ -12,8 +12,7 @@
 namespace polytope {
 
 struct KeyTraits {
-  // typedef uint64_t Key;
-  typedef int64_t Key;
+  typedef int Key;
   static const uint32_t numbits;
   static const uint32_t numbits1d;
   static const Key zero;
